requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
openai
PyGithub
httpx